clang++ -Wmost -Weverything -Wno-c++98-compat

+++ cpp/parhello-c++11/parhello-c++11.cpp
@@ -15,7 +15,7 @@ int main() {
 
   for (auto i : range) {
     futures.emplace_back(
-      std::async(launch::async, [&]() { cout << message[i]; })
+      std::async(launch::async, [&]() { cout << message[(size_t) i]; })
     );
   }
